requirements.txt
fastapi
uvicorn
lyricsgenius
spotipy
python-dotenv
requests
